+++ css/components/admin/dataset/datasets-list-card.scss
@@ -4,7 +4,7 @@
     position: relative;
     padding-right: 20px;
 
-    > .favourite-button {
+    > .c-favourite-button {
       position: absolute;
       top: 0;
       right: 0;
