+++ css/components/app/explore/dataset_widget.scss
@@ -120,7 +120,7 @@
     fill: $charcoal-grey;
   }
 
-  .favourite-button {
+  .c-favourite-button {
     position: absolute;
     right: 0px;
     top: -2px;
