+++ css/components/datasets/list-item.scss
@@ -162,7 +162,7 @@
     fill: $charcoal-grey;
   }
 
-  .favourite-button {
+  .c-favourite-button {
     position: absolute;
     right: 0px;
     top: -2px;
