+++ css/components/ui/favorite-button.scss
@@ -1,0 +1,9 @@
+.c-favourite-button {
+  cursor: pointer;
+
+  &.-loading {
+    opacity: 0.4;
+    user-select: none;
+    cursor: initial;
+  }
+}
