+++ css/components/widgets/widget_card.scss
@@ -22,7 +22,7 @@
     position: relative;
     padding-right: 20px;
 
-    > .favourite-button {
+    .c-favourite-button {
       z-index: 2;
       position: absolute;
       top: 0;
