+++ css/index.scss
@@ -85,6 +85,7 @@
 @import "./components/ui/user-report";
 @import "./components/ui/zoom_control";
 @import "./components/ui/dataset-table";
+@import "./components/ui/favorite-button";
 
 // Modal
 @import "./components/modal/share-modal";
