+++ layout/explore-detail/explore-detail-header/explore-detail-header-component.js
@@ -5,7 +5,7 @@ import PropTypes from 'prop-types';
 import { logEvent } from 'utils/analytics';
 
 // Components
-import ToggleFavorite from 'components/favorites/ToggleFavorite';
+import ToggleFavorite from 'components/favorites/toggle-favorite';
 import Breadcrumbs from 'components/ui/Breadcrumbs';
 import Icon from 'components/ui/Icon';
